chore: Refactor ColumnGrouping into folder with base, utils, and types
