+++ packages/table-core/src/features/column-grouping/ColumnGrouping.types.ts
@@ -1,18 +1,15 @@
-import { RowModel } from '..'
-import { BuiltInAggregationFn, aggregationFns } from '../aggregationFns'
+import { BuiltInAggregationFn } from '../../aggregationFns'
 import {
   AggregationFns,
   Cell,
-  Column,
   ColumnDefTemplate,
   OnChangeFn,
   Row,
   RowData,
+  RowModel,
   Table,
-  TableFeature,
   Updater,
-} from '../types'
-import { isFunction, makeStateUpdater } from '../utils'
+} from '../../types'
 
 export type GroupingState = string[]
 
@@ -235,188 +232,4 @@ export interface GroupingInstance<TData extends RowData> {
    * @link [Guide](https://tanstack.com/table/v8/docs/guide/grouping)
    */
   setGrouping: (updater: Updater<GroupingState>) => void
-}
-
-//
-
-export const ColumnGrouping: TableFeature = {
-  _getDefaultColumnDef: <TData extends RowData>(): GroupingColumnDef<
-    TData,
-    unknown
-  > => {
-    return {
-      aggregatedCell: props => (props.getValue() as any)?.toString?.() ?? null,
-      aggregationFn: 'auto',
-    }
-  },
-
-  _getInitialState: (state): GroupingTableState => {
-    return {
-      grouping: [],
-      ...state,
-    }
-  },
-
-  _getDefaultOptions: <TData extends RowData>(
-    table: Table<TData>
-  ): GroupingOptions => {
-    return {
-      onGroupingChange: makeStateUpdater('grouping', table),
-      groupedColumnMode: 'reorder',
-    }
-  },
-
-  _createColumn: <TData extends RowData, TValue>(
-    column: Column<TData, TValue>,
-    table: Table<TData>
-  ): void => {
-    column.toggleGrouping = () => {
-      table.setGrouping(old => {
-        // Find any existing grouping for this column
-        if (old?.includes(column.id)) {
-          return old.filter(d => d !== column.id)
-        }
-
-        return [...(old ?? []), column.id]
-      })
-    }
-
-    column.getCanGroup = () => {
-      return (
-        (column.columnDef.enableGrouping ?? true) &&
-        (table.options.enableGrouping ?? true) &&
-        (!!column.accessorFn || !!column.columnDef.getGroupingValue)
-      )
-    }
-
-    column.getIsGrouped = () => {
-      return table.getState().grouping?.includes(column.id)
-    }
-
-    column.getGroupedIndex = () => table.getState().grouping?.indexOf(column.id)
-
-    column.getToggleGroupingHandler = () => {
-      const canGroup = column.getCanGroup()
-
-      return () => {
-        if (!canGroup) return
-        column.toggleGrouping()
-      }
-    }
-    column.getAutoAggregationFn = () => {
-      const firstRow = table.getCoreRowModel().flatRows[0]
-
-      const value = firstRow?.getValue(column.id)
-
-      if (typeof value === 'number') {
-        return aggregationFns.sum
-      }
-
-      if (Object.prototype.toString.call(value) === '[object Date]') {
-        return aggregationFns.extent
-      }
-    }
-    column.getAggregationFn = () => {
-      if (!column) {
-        throw new Error()
-      }
-
-      return isFunction(column.columnDef.aggregationFn)
-        ? column.columnDef.aggregationFn
-        : column.columnDef.aggregationFn === 'auto'
-          ? column.getAutoAggregationFn()
-          : table.options.aggregationFns?.[
-              column.columnDef.aggregationFn as string
-            ] ??
-            aggregationFns[
-              column.columnDef.aggregationFn as BuiltInAggregationFn
-            ]
-    }
-  },
-
-  _createTable: <TData extends RowData>(table: Table<TData>): void => {
-    table.setGrouping = updater => table.options.onGroupingChange?.(updater)
-
-    table.resetGrouping = defaultState => {
-      table.setGrouping(defaultState ? [] : table.initialState?.grouping ?? [])
-    }
-
-    table.getPreGroupedRowModel = () => table.getFilteredRowModel()
-    table.getGroupedRowModel = () => {
-      if (!table._getGroupedRowModel && table.options.getGroupedRowModel) {
-        table._getGroupedRowModel = table.options.getGroupedRowModel(table)
-      }
-
-      if (table.options.manualGrouping || !table._getGroupedRowModel) {
-        return table.getPreGroupedRowModel()
-      }
-
-      return table._getGroupedRowModel()
-    }
-  },
-
-  _createRow: <TData extends RowData>(
-    row: Row<TData>,
-    table: Table<TData>
-  ): void => {
-    row.getIsGrouped = () => !!row.groupingColumnId
-    row.getGroupingValue = columnId => {
-      if (row._groupingValuesCache.hasOwnProperty(columnId)) {
-        return row._groupingValuesCache[columnId]
-      }
-
-      const column = table.getColumn(columnId)
-
-      if (!column?.columnDef.getGroupingValue) {
-        return row.getValue(columnId)
-      }
-
-      row._groupingValuesCache[columnId] = column.columnDef.getGroupingValue(
-        row.original
-      )
-
-      return row._groupingValuesCache[columnId]
-    }
-    row._groupingValuesCache = {}
-  },
-
-  _createCell: <TData extends RowData, TValue>(
-    cell: Cell<TData, TValue>,
-    column: Column<TData, TValue>,
-    row: Row<TData>,
-    table: Table<TData>
-  ): void => {
-    const getRenderValue = () =>
-      cell.getValue() ?? table.options.renderFallbackValue
-
-    cell.getIsGrouped = () =>
-      column.getIsGrouped() && column.id === row.groupingColumnId
-    cell.getIsPlaceholder = () => !cell.getIsGrouped() && column.getIsGrouped()
-    cell.getIsAggregated = () =>
-      !cell.getIsGrouped() && !cell.getIsPlaceholder() && !!row.subRows?.length
-  },
-}
-
-export function orderColumns<TData extends RowData>(
-  leafColumns: Column<TData, unknown>[],
-  grouping: string[],
-  groupedColumnMode?: GroupingColumnMode
-) {
-  if (!grouping?.length || !groupedColumnMode) {
-    return leafColumns
-  }
-
-  const nonGroupingColumns = leafColumns.filter(
-    col => !grouping.includes(col.id)
-  )
-
-  if (groupedColumnMode === 'remove') {
-    return nonGroupingColumns
-  }
-
-  const groupingColumns = grouping
-    .map(g => leafColumns.find(col => col.id === g)!)
-    .filter(Boolean)
-
-  return [...groupingColumns, ...nonGroupingColumns]
 }
